# src/config/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1.0.0",
    "title": "Tech Challenge FIAP",
    "description": "System for order management in a restaurant"
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server 🚀"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/customers/": {
      "post": {
        "tags": [
          "Customers"
        ],
        "description": "Create a new customer",
        "parameters": [
          {
            "name": "customerData",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/customers/{id}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/customers/cpf/{cpf}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/customers/email/{email}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Customer": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "123"
        },
        "cpf": {
          "type": "string",
          "example": "123.456.789-00"
        },
        "email": {
          "type": "string",
          "example": "example@example.com"
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}